Move base tests class into a separate file
